Split entity definition file
